requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
pandas==2.0.1
psycopg2-binary==2.9.6
scikit-learn==1.0.2
